Clean up CUDA SFM timing and build gating
Restore the legacy timeSFMBAL benchmark behavior and move the CUDA BAL timing CLI into a CUDA-gated timing/sfm_ba target.

Avoid exposing CUDA headers and cuda-directory sources in CPU-only builds, and add export visibility to public CUDA solver/optimizer APIs.
