+++ gtsam/CMakeLists.txt
@@ -76,6 +76,9 @@ foreach(subdir ${gtsam_subdirs})
     if(GTSAM_ENABLE_CUDA)
       file(GLOB_RECURSE subdir_cuda_srcs "${subdir}/*.cu")
       list(APPEND subdir_srcs ${subdir_cuda_srcs})
+    else()
+      file(GLOB_RECURSE subdir_cuda_srcs "${subdir}/cuda/*")
+      list(REMOVE_ITEM subdir_srcs ${subdir_cuda_srcs})
     endif()
     list(REMOVE_ITEM subdir_srcs ${excluded_sources})
 	file(GLOB subdir_test_files "${subdir}/tests/*")
