+++ gtsam/base/CMakeLists.txt
@@ -5,8 +5,10 @@ install(FILES ${base_headers} DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/gtsam/base
 file(GLOB base_headers_tree "treeTraversal/*.h")
 install(FILES ${base_headers_tree} DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/gtsam/base/treeTraversal)
 
-file(GLOB base_cuda_headers "cuda/*.h")
-install(FILES ${base_cuda_headers} DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/gtsam/base/cuda)
+if(GTSAM_ENABLE_CUDA)
+  file(GLOB base_cuda_headers "cuda/*.h")
+  install(FILES ${base_cuda_headers} DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/gtsam/base/cuda)
+endif()
 
 # Build tests
 add_subdirectory(tests)
