+++ timing/sfm_ba/CMakeLists.txt
@@ -1,0 +1,3 @@
+if(GTSAM_ENABLE_CUDA AND GTSAM_ENABLE_CUDSS)
+  gtsamAddTimingGlob("*.cpp" "" "gtsam" ${GTSAM_BUILD_TIMING_ALWAYS})
+endif()
